Shatter the C++ Comm unit test monolith.
