+++ cpp/test/comm/basic.cpp
@@ -1,6 +1,3 @@
-
-#define BOND_ENABLE_LOG_HANDLER
-
 #include <iostream>
 #include <boost/algorithm/string/replace.hpp>
 
@@ -748,7 +745,8 @@ public:
     }
 };
 
-void init_basic()
+bool init_unit_test()
 {
     InitializeTests<BasicTransportTests, bond::comm::FastWireProtocol>();
+    return true;
 }
