+++ cpp/test/comm/error.cpp
@@ -1,6 +1,3 @@
-
-#define BOND_ENABLE_LOG_HANDLER
-
 #include <iostream>
 #include <boost/algorithm/string/replace.hpp>
 
@@ -315,7 +312,8 @@ public:
 };
 
 
-void init_error()
+bool init_unit_test()
 {
     InitializeTests<ErrorTransportTests, bond::comm::FastWireProtocol>();
+    return true;
 }
