+++ cpp/test/comm/event.cpp
@@ -1,6 +1,3 @@
-
-#define BOND_ENABLE_LOG_HANDLER
-
 #include <iostream>
 #include <boost/algorithm/string/replace.hpp>
 
@@ -155,7 +152,8 @@ public:
 };
 
 
-void init_event()
+bool init_unit_test()
 {
     InitializeTests<EventTransportTests, bond::comm::FastWireProtocol>();
+    return true;
 }
