+++ cpp/test/comm/forward.cpp
@@ -1,6 +1,3 @@
-
-#define BOND_ENABLE_LOG_HANDLER
-
 #include <iostream>
 #include <boost/algorithm/string/replace.hpp>
 
@@ -198,7 +195,8 @@ public:
 };
 
 
-void init_forward()
+bool init_unit_test()
 {
     InitializeTests<ForwardTransportTests, bond::comm::FastWireProtocol>();
+    return true;
 }
