+++ cpp/test/comm/generic_service.cpp
@@ -1,6 +1,3 @@
-
-#define BOND_ENABLE_LOG_HANDLER
-
 #include <iostream>
 #include <boost/algorithm/string/replace.hpp>
 
@@ -268,7 +265,8 @@ public:
 };
 
 
-void init_generic_service()
+bool init_unit_test()
 {
     InitializeTests<GenericServiceTransportTests, bond::comm::FastWireProtocol>();
+    return true;
 }
