+++ cpp/test/comm/jumbo.cpp
@@ -1,6 +1,3 @@
-
-#define BOND_ENABLE_LOG_HANDLER
-
 #include <iostream>
 #include <boost/algorithm/string/replace.hpp>
 
@@ -467,7 +464,8 @@ public:
 };
 
 
-void init_jumbo()
+bool init_unit_test()
 {
     InitializeTests<JumboTransportTests, bond::comm::FastWireProtocol>();
+    return true;
 }
