+++ cpp/test/comm/protocol.cpp
@@ -1,6 +1,3 @@
-
-#define BOND_ENABLE_LOG_HANDLER
-
 #include <iostream>
 #include <boost/algorithm/string/replace.hpp>
 
@@ -245,7 +242,8 @@ public:
     }
 };
 
-void init_protocol()
+bool init_unit_test()
 {
     InitializeTests<ProtocolTransportTests, bond::comm::FastWireProtocol>();
+    return true;
 }
