+++ cpp/test/comm/session.cpp
@@ -1,6 +1,3 @@
-
-#define BOND_ENABLE_LOG_HANDLER
-
 #include <iostream>
 #include <boost/algorithm/string/replace.hpp>
 
@@ -187,7 +184,8 @@ public:
 };
 
 
-void init_session()
+bool init_unit_test()
 {
     InitializeTests<SessionTransportTests, bond::comm::FastWireProtocol>();
+    return true;
 }
