+++ cpp/test/comm/timeout.cpp
@@ -1,6 +1,3 @@
-
-#define BOND_ENABLE_LOG_HANDLER
-
 #include <iostream>
 #include <boost/algorithm/string/replace.hpp>
 
@@ -336,7 +333,8 @@ public:
 };
 
 
-void init_timeout()
+bool init_unit_test()
 {
     InitializeTests<TimeoutTransportTests, bond::comm::FastWireProtocol>();
+    return true;
 }
