+++ cpp/test/comm/unexpected.cpp
@@ -1,6 +1,3 @@
-
-#define BOND_ENABLE_LOG_HANDLER
-
 #include <iostream>
 #include <boost/algorithm/string/replace.hpp>
 
@@ -281,7 +278,8 @@ public:
 };
 
 
-void init_unexpected()
+bool init_unit_test()
 {
     InitializeTests<UnexpectedTransportTests, bond::comm::FastWireProtocol>();
+    return true;
 }
