+++ cpp/test/comm/void.cpp
@@ -1,6 +1,3 @@
-
-#define BOND_ENABLE_LOG_HANDLER
-
 #include <iostream>
 #include <boost/algorithm/string/replace.hpp>
 
@@ -338,7 +335,8 @@ public:
 };
 
 
-void init_void()
+bool init_unit_test()
 {
     InitializeTests<VoidTransportTests, bond::comm::FastWireProtocol>();
+    return true;
 }
