+++ examples/cpp/comm/error/CMakeLists.txt
@@ -1,3 +1,3 @@
-add_bond_test (error error.bond error.cpp COMM)
+add_bond_test (error_example error.bond error.cpp COMM)
 
 target_compile_definitions (error PRIVATE -D_WIN32_WINNT=0x0600)
