+++ examples/cpp/comm/generic/CMakeLists.txt
@@ -1,3 +1,3 @@
-add_bond_test (generic generic.bond generic.cpp COMM)
+add_bond_test (generic_example generic.bond generic.cpp COMM)
 
-target_compile_definitions (generic PRIVATE -D_WIN32_WINNT=0x0600)
+target_compile_definitions (generic_example PRIVATE -D_WIN32_WINNT=0x0600)
