+++ examples/cpp/comm/void/CMakeLists.txt
@@ -1,3 +1,3 @@
-add_bond_test (void void.bond void.cpp COMM)
+add_bond_test (void_example void.bond void.cpp COMM)
 
-target_compile_definitions (void PRIVATE -D_WIN32_WINNT=0x0600)
+target_compile_definitions (void_example PRIVATE -D_WIN32_WINNT=0x0600)
